Split up headers
