+++ common/components/Header/NewHeader/NewHeader.scss
@@ -1,125 +1,15 @@
-$header-background-color: #163150;
-$header-icon-color: #1eb8e7;
-$dropdown-background-color: #0f253f;
+.mobile-only-header {
+  display: block;
 
-.NewHeader {
-  display: flex;
-  flex-direction: column;
-  background: $header-background-color;
-  color: #fff;
-
-  i {
-    color: $header-icon-color;
+  @media (min-width: 1000px) {
+    display: none;
   }
-  &-top {
-    display: flex;
-    align-items: center;
-    padding: 12px 50px;
-    border-bottom: 1px solid #3e546d;
-    font-size: 14px;
-    text-transform: uppercase;
+}
 
-    ul > li:first-of-type {
-      margin-right: 38px;
-    }
-    &-links {
-      margin: 0;
-      padding: 0;
-      list-style-type: none;
+.desktop-only-header {
+  display: none;
 
-      li {
-        display: inline;
-        font-weight: bold;
-        letter-spacing: 1px;
-        cursor: pointer;
-
-        a {
-          color: #fff;
-        }
-      }
-    }
-    &-left {
-      flex: 1;
-    }
-    &-center {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-
-      img {
-        width: 160px;
-        height: 39px;
-      }
-    }
-    &-right {
-      flex: 1;
-      display: flex;
-      justify-content: flex-end;
-    }
-  }
-  &-bottom {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    font-size: 14px;
-    text-transform: uppercase;
-    letter-spacing: 1px;
-
-    &-links {
-      margin: 0;
-      padding: 0;
-      list-style-type: none;
-
-      & > li {
-        display: inline-flex;
-        height: 58px;
-        margin: 0;
-        font-weight: bold;
-        letter-spacing: 1px;
-      }
-      &-item {
-        position: relative;
-        display: inline-flex;
-        align-items: center;
-        justify-content: center;
-        padding: 0 22px;
-        height: 58px;
-        cursor: pointer;
-
-        a {
-          color: #fff;
-        }
-        i {
-          margin-left: 5px;
-
-          &.create-icon {
-            margin-right: 5px;
-            margin-left: 0;
-          }
-        }
-      }
-      &-dropdown {
-        position: absolute;
-        z-index: 200;
-        top: 100%;
-        width: 200px;
-        margin: 0;
-        padding: 8px 16px;
-        background: $dropdown-background-color;
-        text-transform: initial;
-        list-style-type: none;
-        cursor: initial;
-
-        li {
-          font-size: 12px;
-          padding: 8px 0;
-
-          a {
-            color: #fff;
-            font-weight: normal;
-          }
-        }
-      }
-    }
+  @media (min-width: 1000px) {
+    display: block;
   }
 }
