+++ common/components/Header/NewHeader/components/index.ts
@@ -1,0 +1,2 @@
+export { default as DesktopHeader } from './DesktopHeader';
+export { default as MobileHeader } from './MobileHeader';
