+++ common/components/Header/NewHeader/constants.ts
@@ -1,0 +1,52 @@
+export const LINKSET = {
+  SEND_AND_RECEIVE: [
+    {
+      to: '/account',
+      title: 'Send Ether & Tokens'
+    },
+    {
+      to: '/account/request',
+      title: 'Request Payment'
+    },
+    {
+      to: '/account/info',
+      title: 'Wallet Info'
+    },
+    {
+      to: '/account/recent-txs',
+      title: 'View Recent Transactions'
+    },
+    {
+      to: '/account/address-book',
+      title: 'Address Book'
+    }
+  ],
+  BUY_AND_EXCHANGE: [
+    {
+      to: '/swap',
+      title: 'Swap ETH & Tokens'
+    }
+  ],
+  TOOLS: [
+    {
+      to: '/sign-and-verify-message',
+      title: 'Sign & Verify Message'
+    },
+    {
+      to: '/contracts',
+      title: 'Interact with Contracts'
+    },
+    {
+      to: '/tx-status',
+      title: 'Check Transaction Status'
+    },
+    {
+      to: '/pushTx',
+      title: 'Broadcast Transaction'
+    },
+    {
+      to: '/ens',
+      title: 'ENS Domains'
+    }
+  ]
+};
